renamed auth reducer
