+++ src/reducer/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import auth from './auth';
+import auth from './auth-reducer';
 import game from './game';
 
 export default combineReducers({
